requirements.txt
aiohttp
backtesting
beautifulsoup4
crewai
instructor
joblib
matplotlib
mplfinance
numpy
openai
pandas
pydantic
pydantic-settings
pytest
python-binance
python-dotenv
PyYAML
requests
scikit-learn
scipy
seaborn
ta-lib
websockets
psutil
